+++ tests/units/base.js
@@ -87,6 +87,9 @@ test( 'Properties', function() {
 })
 
 
+
+
+
 module( 'Formatting setup', {
     setup: function() {
         $DOM.append( $INPUT.clone().attr( 'name', 'picker' ) )
@@ -104,6 +107,30 @@ module( 'Formatting setup', {
 test( 'Hidden suffix', function() {
     var picker = this.picker
     strictEqual( picker.$node[0].name + '_submit', picker._hidden.name, 'Correct hidden element `name` suffix' )
+})
+
+
+
+
+
+module( 'Container setup', {
+    setup: function() {
+        $DOM.append( $INPUT.clone().attr( 'name', 'picker' ), $( '<div id="outlet"/>' ) )
+        var $input = $DOM.find( 'input' ).pickadate({
+            formatSubmit: 'yyyy/mm/dd',
+            container: '#outlet'
+        })
+        this.picker = $input.pickadate( 'picker' )
+    },
+    teardown: function() {
+        this.picker.stop()
+        $DOM.empty()
+    }
+})
+
+test( 'Picker root outlet', function() {
+    var picker = this.picker
+    strictEqual( picker.$root[0].parentNode.id, 'outlet', 'Correct root outlet' )
 })